refactor: organize icons
