+++ src/components/content/article-share-buttons.tsx
@@ -3,7 +3,7 @@
 import { cn } from '@/lib/utils';
 
 import { Button } from '@/components/shadcn-ui/button';
-import { BrandIcons } from '@/components/icons/brand';
+import { BrandIcons } from '@/components/icons/brand-icons';
 
 interface ArticleShareButtonsProps {
   url: string;
