+++ src/components/content/mdx-components.tsx
@@ -4,7 +4,7 @@ import { cn } from '@/lib/utils';
 
 import { Button } from '@/components/shadcn-ui/button';
 import { LinkPreview } from '@/components/content/link-preview';
-import { getIconForLanguageExtension } from '@/components/icons/additional-icons';
+import { getIconForLanguageExtension } from '@/components/icons/brand-icons';
 import { Callout } from '@/components/shared/callout';
 import { CopyButton } from '@/components/shared/copy-button';
