+++ src/components/icons/brand-icons.tsx
@@ -2,243 +2,71 @@ import { FileIcon } from 'lucide-react';
 
 type IconProps = React.HTMLAttributes<SVGElement>;
 
-export const AdditionalIcons = {
-  aumon: (props: IconProps) => (
+export const BrandIcons = {
+  facebook: (props: IconProps) => (
     <svg
+      role="img"
+      viewBox="0 0 24 24"
       xmlns="http://www.w3.org/2000/svg"
-      width="16"
-      height="16"
-      viewBox="0 0 1200 600"
-      fill="none"
       {...props}
     >
-      <g clipPath="url(#clip0_189_51759)">
-        <g clipPath="url(#clip1_189_51759)">
-          <path
-            d="M351.387 531.689C283.892 449.197 198.934 535.203 298.727 612.537C398.521 689.872 498.407 698.395 567.776 707.523C637.145 716.651 650.938 685.824 585.354 670.044C519.77 654.264 418.882 614.181 351.387 531.689Z"
-            fill="#282C32"
-            stroke="#282C32"
-            strokeWidth="16"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-          <path
-            d="M894.854 44.6386C936.866 101.01 905.411 170.929 845.891 186.926C775.094 205.943 722.635 175.586 663.629 227.461C682.909 125.324 706.858 56.2479 767.602 22.1234C810.224 -1.82612 867.023 7.26755 894.854 44.6386ZM791.759 117.662C801.484 131.486 821.046 134.487 835.458 124.344C849.866 114.221 853.656 94.787 843.932 80.9633C834.207 67.1395 814.645 64.1387 800.233 74.2814C785.825 84.4046 782.034 103.839 791.759 117.662Z"
-            fill="#E93F34"
-            stroke="#282C32"
-            strokeWidth="16"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-          <g clipPath="url(#clip2_189_51759)">
-            <path
-              d="M561.332 253.132C572.837 255.304 584.625 244.024 587.662 227.939C590.698 211.853 583.833 197.052 572.328 194.88C560.823 192.709 549.034 203.988 545.998 220.074C542.961 236.159 549.827 250.96 561.332 253.132Z"
-              fill="#282C32"
-              stroke="#282C32"
-              strokeWidth="16"
-              strokeLinecap="round"
-              strokeLinejoin="round"
-            />
-            <path
-              d="M749.547 288.661C761.053 290.833 772.841 279.553 775.878 263.468C778.914 247.382 772.049 232.582 760.544 230.41C749.038 228.238 737.25 239.517 734.213 255.603C731.177 271.689 738.042 286.489 749.547 288.661Z"
-              fill="#282C32"
-              stroke="#282C32"
-              strokeWidth="16"
-              strokeLinecap="round"
-              strokeLinejoin="round"
-            />
-          </g>
-          <path
-            d="M572.637 709.631C638.611 722.085 651.835 768.157 700.495 777.343C724.295 781.836 797.039 785.167 831.114 672.696C860.969 574.164 824.324 239.539 666.963 209.834C509.602 180.129 353.495 478.348 345.372 581.002C336.104 698.136 405.053 721.572 428.853 726.065C477.533 735.254 506.642 697.173 572.597 709.623L572.637 709.631Z"
-            fill="#FAD62C"
-            stroke="#282C32"
-            strokeWidth="16"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-          <g clipPath="url(#clip3_189_51759)">
-            <g clipPath="url(#clip4_189_51759)">
-              <path
-                d="M613.922 399.13C636.206 403.337 657.809 387.996 662.176 364.866C666.542 341.737 652.017 319.576 629.734 315.37C607.451 311.163 585.847 326.504 581.481 349.634C577.115 372.764 591.639 394.924 613.922 399.13Z"
-                fill="white"
-              />
-              <path
-                d="M714.341 429.725C723.023 391.542 684.465 350.222 628.22 337.433C571.975 324.644 519.342 345.23 510.66 383.412C501.978 421.595 540.536 462.915 596.781 475.704C653.026 488.493 705.659 467.907 714.341 429.725Z"
-                fill="white"
-              />
-              <path
-                d="M520.55 386.471C532.066 388.645 542.497 384.608 543.847 377.456C545.197 370.303 536.956 362.742 525.44 360.568C513.924 358.394 503.494 362.431 502.143 369.583C500.793 376.736 509.034 384.297 520.55 386.471Z"
-                fill="#F9A621"
-              />
-              <path
-                d="M705.484 421.381C717 423.555 727.43 419.518 728.781 412.366C730.131 405.213 721.89 397.652 710.374 395.478C698.858 393.304 688.427 397.34 687.077 404.493C685.727 411.646 693.968 419.207 705.484 421.381Z"
-                fill="#F9A621"
-              />
-            </g>
-          </g>
-          <path
-            d="M697.296 340.223C679.634 339.657 664.143 342.473 656.092 349.542C674.91 354.336 686.19 366.174 688.554 368.309"
-            stroke="#282C32"
-            strokeWidth="16"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-          <path
-            d="M557.76 313.883C574.413 319.795 587.812 328.064 592.732 337.581C573.462 335.185 558.642 342.096 555.662 343.223"
-            stroke="#282C32"
-            strokeWidth="16"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-          <path
-            d="M612.787 397.379L617.602 398.288L634.435 421.575L619.641 439.888L594.78 435.196L587.681 412.749L612.787 397.379Z"
-            fill="#DD938E"
-          />
-          <path
-            d="M607.672 437.629C625.164 440.931 629.705 430.676 629.705 430.676C629.705 430.676 623.668 416.51 612.072 414.321C600.457 412.128 589.691 423.122 589.691 423.122C589.691 423.122 590.181 434.328 607.672 437.629Z"
-            fill="#F9BDB4"
-          />
-          <path
-            d="M588.633 408.675C584.563 424.413 586.388 433.612 607.692 437.633"
-            stroke="#282C32"
-            strokeWidth="12"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-          <path
-            d="M635.034 417.434C633.086 433.573 628.033 441.473 606.729 437.451"
-            stroke="#282C32"
-            strokeWidth="12"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-          <path
-            d="M616.431 391.229C614.372 402.136 583.105 418.765 569.467 396.02"
-            stroke="#282C32"
-            strokeWidth="12"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-          <path
-            d="M616.431 391.229C614.372 402.136 637.426 429.019 658.417 412.811"
-            stroke="#282C32"
-            strokeWidth="12"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-          <path
-            d="M632.122 382.935C630.991 388.929 623.959 392.65 616.432 391.229C608.905 389.808 603.713 383.78 604.844 377.786C605.976 371.792 612.451 371.02 619.978 372.441C627.505 373.861 633.257 376.922 632.122 382.935Z"
-            fill="#282C32"
-            stroke="#282C32"
-            strokeWidth="12"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-          <path
-            d="M584.542 646.418C606.293 650.524 609.792 663.604 627.509 666.949C645.226 670.293 662.374 665.389 668.31 633.944C674.246 602.499 649.698 506.276 604.467 497.738C559.236 489.2 516.651 574.786 511.086 604.265C505.521 633.744 518.347 646.342 537.047 649.872C555.747 653.402 562.79 642.312 584.542 646.418Z"
-            fill="white"
-          />
-          <circle
-            cx="466.25"
-            cy="448.24"
-            r="41"
-            transform="rotate(10.6899 466.25 448.24)"
-            fill="#FAD62C"
-          />
-          <path
-            d="M482.142 401.044C472.96 392.859 464.926 382.265 459.818 372.773C445.334 345.777 437.674 315.735 408.833 327.815C379.488 340.084 377.916 379.354 408.63 432.718C410.841 436.534 416.865 451.308 418.355 460.341"
-            fill="#FAD62C"
-          />
-          <path
-            d="M482.142 401.044C472.96 392.859 464.926 382.265 459.818 372.773C445.334 345.777 437.674 315.735 408.833 327.815C379.488 340.084 377.916 379.354 408.63 432.718C410.841 436.534 416.865 451.308 418.355 460.341"
-            stroke="#282C32"
-            strokeWidth="16"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-          <circle
-            cx="41"
-            cy="41"
-            r="41"
-            transform="matrix(-0.982645 -0.185494 -0.185494 0.982645 814.833 472.318)"
-            fill="#FAD62C"
-          />
-          <path
-            d="M755.927 550.875C791.444 563.38 805.903 572.175 820.823 582.95C835.744 593.724 859.091 593.389 871.778 577.934C884.465 562.479 882.218 539.667 866.763 526.98L808.937 481.382"
-            fill="#FAD62C"
-          />
-          <path
-            d="M755.927 550.875C791.444 563.38 805.903 572.175 820.823 582.95C835.744 593.724 859.091 593.389 871.778 577.934C884.465 562.479 882.218 539.667 866.763 526.98L808.937 481.382"
-            stroke="#282C32"
-            strokeWidth="16"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-        </g>
-      </g>
-      <defs>
-        <clipPath id="clip0_189_51759">
-          <rect width="1200" height="600" fill="white" />
-        </clipPath>
-        <clipPath id="clip1_189_51759">
-          <rect
-            width="1024"
-            height="1024"
-            fill="white"
-            transform="translate(232.753 -250.193) rotate(10.6899)"
-          />
-        </clipPath>
-        <clipPath id="clip2_189_51759">
-          <rect
-            width="249.92"
-            height="75.28"
-            fill="white"
-            transform="translate(545.118 181.603) rotate(10.6899)"
-          />
-        </clipPath>
-        <clipPath id="clip3_189_51759">
-          <rect
-            width="256"
-            height="256"
-            fill="white"
-            transform="translate(512.509 246.317) rotate(10.6899)"
-          />
-        </clipPath>
-        <clipPath id="clip4_189_51759">
-          <rect
-            width="230.64"
-            height="163.78"
-            fill="white"
-            transform="translate(516.415 293.978) rotate(10.6899)"
-          />
-        </clipPath>
-      </defs>
+      <title>Facebook</title>
+      <path
+        fill="currentColor"
+        d="M9.101 23.691v-7.98H6.627v-3.667h2.474v-1.58c0-4.085 1.848-5.978 5.858-5.978.401 0 .955.042 1.468.103a8.68 8.68 0 0 1 1.141.195v3.325a8.623 8.623 0 0 0-.653-.036 26.805 26.805 0 0 0-.733-.009c-.707 0-1.259.096-1.675.309a1.686 1.686 0 0 0-.679.622c-.258.42-.374.995-.374 1.752v1.297h3.919l-.386 2.103-.287 1.564h-3.246v8.245C19.396 23.238 24 18.179 24 12.044c0-6.627-5.373-12-12-12s-12 5.373-12 12c0 5.628 3.874 10.35 9.101 11.647Z"
+      />
     </svg>
   ),
-  logo: (props: IconProps) => (
-    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 256 256" {...props}>
-      <rect width="256" height="256" fill="none" />
-      <line
-        x1="208"
-        y1="128"
-        x2="128"
-        y2="208"
-        fill="none"
-        stroke="currentColor"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-        strokeWidth="32"
+  instagram: (props: IconProps) => (
+    <svg
+      role="img"
+      viewBox="0 0 24 24"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <title>Instagram</title>
+      <path d="M7.0301.084c-1.2768.0602-2.1487.264-2.911.5634-.7888.3075-1.4575.72-2.1228 1.3877-.6652.6677-1.075 1.3368-1.3802 2.127-.2954.7638-.4956 1.6365-.552 2.914-.0564 1.2775-.0689 1.6882-.0626 4.947.0062 3.2586.0206 3.6671.0825 4.9473.061 1.2765.264 2.1482.5635 2.9107.308.7889.72 1.4573 1.388 2.1228.6679.6655 1.3365 1.0743 2.1285 1.38.7632.295 1.6361.4961 2.9134.552 1.2773.056 1.6884.069 4.9462.0627 3.2578-.0062 3.668-.0207 4.9478-.0814 1.28-.0607 2.147-.2652 2.9098-.5633.7889-.3086 1.4578-.72 2.1228-1.3881.665-.6682 1.0745-1.3378 1.3795-2.1284.2957-.7632.4966-1.636.552-2.9124.056-1.2809.0692-1.6898.063-4.948-.0063-3.2583-.021-3.6668-.0817-4.9465-.0607-1.2797-.264-2.1487-.5633-2.9117-.3084-.7889-.72-1.4568-1.3876-2.1228C21.2982 1.33 20.628.9208 19.8378.6165 19.074.321 18.2017.1197 16.9244.0645 15.6471.0093 15.236-.005 11.977.0014 8.718.0076 8.31.0215 7.0301.0839m.1402 21.6932c-1.17-.0509-1.8053-.2453-2.2287-.408-.5606-.216-.96-.4771-1.3819-.895-.422-.4178-.6811-.8186-.9-1.378-.1644-.4234-.3624-1.058-.4171-2.228-.0595-1.2645-.072-1.6442-.079-4.848-.007-3.2037.0053-3.583.0607-4.848.05-1.169.2456-1.805.408-2.2282.216-.5613.4762-.96.895-1.3816.4188-.4217.8184-.6814 1.3783-.9003.423-.1651 1.0575-.3614 2.227-.4171 1.2655-.06 1.6447-.072 4.848-.079 3.2033-.007 3.5835.005 4.8495.0608 1.169.0508 1.8053.2445 2.228.408.5608.216.96.4754 1.3816.895.4217.4194.6816.8176.9005 1.3787.1653.4217.3617 1.056.4169 2.2263.0602 1.2655.0739 1.645.0796 4.848.0058 3.203-.0055 3.5834-.061 4.848-.051 1.17-.245 1.8055-.408 2.2294-.216.5604-.4763.96-.8954 1.3814-.419.4215-.8181.6811-1.3783.9-.4224.1649-1.0577.3617-2.2262.4174-1.2656.0595-1.6448.072-4.8493.079-3.2045.007-3.5825-.006-4.848-.0608M16.953 5.5864A1.44 1.44 0 1 0 18.39 4.144a1.44 1.44 0 0 0-1.437 1.4424M5.8385 12.012c.0067 3.4032 2.7706 6.1557 6.173 6.1493 3.4026-.0065 6.157-2.7701 6.1506-6.1733-.0065-3.4032-2.771-6.1565-6.174-6.1498-3.403.0067-6.156 2.771-6.1496 6.1738M8 12.0077a4 4 0 1 1 4.008 3.9921A3.9996 3.9996 0 0 1 8 12.0077" />
+    </svg>
+  ),
+  line: (props: IconProps) => (
+    <svg
+      role="img"
+      viewBox="0 0 24 24"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <title>LINE</title>
+      <path
+        fill="currentColor"
+        d="M19.365 9.863c.349 0 .63.285.63.631 0 .345-.281.63-.63.63H17.61v1.125h1.755c.349 0 .63.283.63.63 0 .344-.281.629-.63.629h-2.386c-.345 0-.627-.285-.627-.629V8.108c0-.345.282-.63.63-.63h2.386c.346 0 .627.285.627.63 0 .349-.281.63-.63.63H17.61v1.125h1.755zm-3.855 3.016c0 .27-.174.51-.432.596-.064.021-.133.031-.199.031-.211 0-.391-.09-.51-.25l-2.443-3.317v2.94c0 .344-.279.629-.631.629-.346 0-.626-.285-.626-.629V8.108c0-.27.173-.51.43-.595.06-.023.136-.033.194-.033.195 0 .375.104.495.254l2.462 3.33V8.108c0-.345.282-.63.63-.63.345 0 .63.285.63.63v4.771zm-5.741 0c0 .344-.282.629-.631.629-.345 0-.627-.285-.627-.629V8.108c0-.345.282-.63.63-.63.346 0 .628.285.628.63v4.771zm-2.466.629H4.917c-.345 0-.63-.285-.63-.629V8.108c0-.345.285-.63.63-.63.348 0 .63.285.63.63v4.141h1.756c.348 0 .629.283.629.63 0 .344-.282.629-.629.629M24 10.314C24 4.943 18.615.572 12 .572S0 4.943 0 10.314c0 4.811 4.27 8.842 10.035 9.608.391.082.923.258 1.058.59.12.301.079.766.038 1.08l-.164 1.02c-.045.301-.24 1.186 1.049.645 1.291-.539 6.916-4.078 9.436-6.975C23.176 14.393 24 12.458 24 10.314"
       />
-      <line
-        x1="192"
-        y1="40"
-        x2="40"
-        y2="192"
-        fill="none"
-        stroke="currentColor"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-        strokeWidth="32"
+    </svg>
+  ),
+  pinterest: (props: IconProps) => (
+    <svg
+      role="img"
+      viewBox="0 0 24 24"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <title>Pinterest</title>
+      <path
+        fill="currentColor"
+        d="M12.017 0C5.396 0 .029 5.367.029 11.987c0 5.079 3.158 9.417 7.618 11.162-.105-.949-.199-2.403.041-3.439.219-.937 1.406-5.957 1.406-5.957s-.359-.72-.359-1.781c0-1.663.967-2.911 2.168-2.911 1.024 0 1.518.769 1.518 1.688 0 1.029-.653 2.567-.992 3.992-.285 1.193.6 2.165 1.775 2.165 2.128 0 3.768-2.245 3.768-5.487 0-2.861-2.063-4.869-5.008-4.869-3.41 0-5.409 2.562-5.409 5.199 0 1.033.394 2.143.889 2.741.099.12.112.225.085.345-.09.375-.293 1.199-.334 1.363-.053.225-.172.271-.401.165-1.495-.69-2.433-2.878-2.433-4.646 0-3.776 2.748-7.252 7.92-7.252 4.158 0 7.392 2.967 7.392 6.923 0 4.135-2.607 7.462-6.233 7.462-1.214 0-2.354-.629-2.758-1.379l-.749 2.848c-.269 1.045-1.004 2.352-1.498 3.146 1.123.345 2.306.535 3.55.535 6.607 0 11.985-5.365 11.985-11.987C23.97 5.39 18.592.026 11.985.026L12.017 0z"
+      />
+    </svg>
+  ),
+  x: (props: IconProps) => (
+    <svg
+      role="img"
+      viewBox="0 0 24 24"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <title>X</title>
+      <path
+        fill="currentColor"
+        d="M18.901 1.153h3.68l-8.04 9.19L24 22.846h-7.406l-5.8-7.584-6.638 7.584H.474l8.6-9.83L0 1.154h7.594l5.243 6.932ZM17.61 20.644h2.039L6.486 3.24H4.298Z"
       />
     </svg>
   ),
@@ -343,22 +171,6 @@ export const AdditionalIcons = {
       />
     </svg>
   ),
-  spinner: (props: IconProps) => (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      {...props}
-    >
-      <path d="M21 12a9 9 0 1 1-6.219-8.56" />
-    </svg>
-  ),
   json: (props: IconProps) => (
     <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" {...props}>
       <path
@@ -404,15 +216,15 @@ export const AdditionalIcons = {
 export function getIconForLanguageExtension(language: string) {
   switch (language) {
     case 'json':
-      return <AdditionalIcons.json />;
+      return <BrandIcons.json />;
     case 'css':
-      return <AdditionalIcons.css className="fill-foreground" />;
+      return <BrandIcons.css className="fill-foreground" />;
     case 'js':
     case 'jsx':
     case 'ts':
     case 'tsx':
     case 'typescript':
-      return <AdditionalIcons.ts className="fill-foreground" />;
+      return <BrandIcons.ts className="fill-foreground" />;
     default:
       return <FileIcon />;
   }
