--- src/components/icons/brand.tsx
@@ -1,102 +1,0 @@
-import { IconProps } from '@/components/icons';
-
-export const BrandIcons = {
-  facebook: (props: IconProps) => (
-    <svg
-      role="img"
-      viewBox="0 0 24 24"
-      xmlns="http://www.w3.org/2000/svg"
-      {...props}
-    >
-      <title>Facebook</title>
-      <path
-        fill="currentColor"
-        d="M9.101 23.691v-7.98H6.627v-3.667h2.474v-1.58c0-4.085 1.848-5.978 5.858-5.978.401 0 .955.042 1.468.103a8.68 8.68 0 0 1 1.141.195v3.325a8.623 8.623 0 0 0-.653-.036 26.805 26.805 0 0 0-.733-.009c-.707 0-1.259.096-1.675.309a1.686 1.686 0 0 0-.679.622c-.258.42-.374.995-.374 1.752v1.297h3.919l-.386 2.103-.287 1.564h-3.246v8.245C19.396 23.238 24 18.179 24 12.044c0-6.627-5.373-12-12-12s-12 5.373-12 12c0 5.628 3.874 10.35 9.101 11.647Z"
-      />
-    </svg>
-  ),
-  instagram: (props: IconProps) => (
-    <svg
-      role="img"
-      viewBox="0 0 24 24"
-      xmlns="http://www.w3.org/2000/svg"
-      {...props}
-    >
-      <title>Instagram</title>
-      <path d="M7.0301.084c-1.2768.0602-2.1487.264-2.911.5634-.7888.3075-1.4575.72-2.1228 1.3877-.6652.6677-1.075 1.3368-1.3802 2.127-.2954.7638-.4956 1.6365-.552 2.914-.0564 1.2775-.0689 1.6882-.0626 4.947.0062 3.2586.0206 3.6671.0825 4.9473.061 1.2765.264 2.1482.5635 2.9107.308.7889.72 1.4573 1.388 2.1228.6679.6655 1.3365 1.0743 2.1285 1.38.7632.295 1.6361.4961 2.9134.552 1.2773.056 1.6884.069 4.9462.0627 3.2578-.0062 3.668-.0207 4.9478-.0814 1.28-.0607 2.147-.2652 2.9098-.5633.7889-.3086 1.4578-.72 2.1228-1.3881.665-.6682 1.0745-1.3378 1.3795-2.1284.2957-.7632.4966-1.636.552-2.9124.056-1.2809.0692-1.6898.063-4.948-.0063-3.2583-.021-3.6668-.0817-4.9465-.0607-1.2797-.264-2.1487-.5633-2.9117-.3084-.7889-.72-1.4568-1.3876-2.1228C21.2982 1.33 20.628.9208 19.8378.6165 19.074.321 18.2017.1197 16.9244.0645 15.6471.0093 15.236-.005 11.977.0014 8.718.0076 8.31.0215 7.0301.0839m.1402 21.6932c-1.17-.0509-1.8053-.2453-2.2287-.408-.5606-.216-.96-.4771-1.3819-.895-.422-.4178-.6811-.8186-.9-1.378-.1644-.4234-.3624-1.058-.4171-2.228-.0595-1.2645-.072-1.6442-.079-4.848-.007-3.2037.0053-3.583.0607-4.848.05-1.169.2456-1.805.408-2.2282.216-.5613.4762-.96.895-1.3816.4188-.4217.8184-.6814 1.3783-.9003.423-.1651 1.0575-.3614 2.227-.4171 1.2655-.06 1.6447-.072 4.848-.079 3.2033-.007 3.5835.005 4.8495.0608 1.169.0508 1.8053.2445 2.228.408.5608.216.96.4754 1.3816.895.4217.4194.6816.8176.9005 1.3787.1653.4217.3617 1.056.4169 2.2263.0602 1.2655.0739 1.645.0796 4.848.0058 3.203-.0055 3.5834-.061 4.848-.051 1.17-.245 1.8055-.408 2.2294-.216.5604-.4763.96-.8954 1.3814-.419.4215-.8181.6811-1.3783.9-.4224.1649-1.0577.3617-2.2262.4174-1.2656.0595-1.6448.072-4.8493.079-3.2045.007-3.5825-.006-4.848-.0608M16.953 5.5864A1.44 1.44 0 1 0 18.39 4.144a1.44 1.44 0 0 0-1.437 1.4424M5.8385 12.012c.0067 3.4032 2.7706 6.1557 6.173 6.1493 3.4026-.0065 6.157-2.7701 6.1506-6.1733-.0065-3.4032-2.771-6.1565-6.174-6.1498-3.403.0067-6.156 2.771-6.1496 6.1738M8 12.0077a4 4 0 1 1 4.008 3.9921A3.9996 3.9996 0 0 1 8 12.0077" />
-    </svg>
-  ),
-  line: (props: IconProps) => (
-    <svg
-      role="img"
-      viewBox="0 0 24 24"
-      xmlns="http://www.w3.org/2000/svg"
-      {...props}
-    >
-      <title>LINE</title>
-      <path
-        fill="currentColor"
-        d="M19.365 9.863c.349 0 .63.285.63.631 0 .345-.281.63-.63.63H17.61v1.125h1.755c.349 0 .63.283.63.63 0 .344-.281.629-.63.629h-2.386c-.345 0-.627-.285-.627-.629V8.108c0-.345.282-.63.63-.63h2.386c.346 0 .627.285.627.63 0 .349-.281.63-.63.63H17.61v1.125h1.755zm-3.855 3.016c0 .27-.174.51-.432.596-.064.021-.133.031-.199.031-.211 0-.391-.09-.51-.25l-2.443-3.317v2.94c0 .344-.279.629-.631.629-.346 0-.626-.285-.626-.629V8.108c0-.27.173-.51.43-.595.06-.023.136-.033.194-.033.195 0 .375.104.495.254l2.462 3.33V8.108c0-.345.282-.63.63-.63.345 0 .63.285.63.63v4.771zm-5.741 0c0 .344-.282.629-.631.629-.345 0-.627-.285-.627-.629V8.108c0-.345.282-.63.63-.63.346 0 .628.285.628.63v4.771zm-2.466.629H4.917c-.345 0-.63-.285-.63-.629V8.108c0-.345.285-.63.63-.63.348 0 .63.285.63.63v4.141h1.756c.348 0 .629.283.629.63 0 .344-.282.629-.629.629M24 10.314C24 4.943 18.615.572 12 .572S0 4.943 0 10.314c0 4.811 4.27 8.842 10.035 9.608.391.082.923.258 1.058.59.12.301.079.766.038 1.08l-.164 1.02c-.045.301-.24 1.186 1.049.645 1.291-.539 6.916-4.078 9.436-6.975C23.176 14.393 24 12.458 24 10.314"
-      />
-    </svg>
-  ),
-  pinterest: (props: IconProps) => (
-    <svg
-      role="img"
-      viewBox="0 0 24 24"
-      xmlns="http://www.w3.org/2000/svg"
-      {...props}
-    >
-      <title>Pinterest</title>
-      <path
-        fill="currentColor"
-        d="M12.017 0C5.396 0 .029 5.367.029 11.987c0 5.079 3.158 9.417 7.618 11.162-.105-.949-.199-2.403.041-3.439.219-.937 1.406-5.957 1.406-5.957s-.359-.72-.359-1.781c0-1.663.967-2.911 2.168-2.911 1.024 0 1.518.769 1.518 1.688 0 1.029-.653 2.567-.992 3.992-.285 1.193.6 2.165 1.775 2.165 2.128 0 3.768-2.245 3.768-5.487 0-2.861-2.063-4.869-5.008-4.869-3.41 0-5.409 2.562-5.409 5.199 0 1.033.394 2.143.889 2.741.099.12.112.225.085.345-.09.375-.293 1.199-.334 1.363-.053.225-.172.271-.401.165-1.495-.69-2.433-2.878-2.433-4.646 0-3.776 2.748-7.252 7.92-7.252 4.158 0 7.392 2.967 7.392 6.923 0 4.135-2.607 7.462-6.233 7.462-1.214 0-2.354-.629-2.758-1.379l-.749 2.848c-.269 1.045-1.004 2.352-1.498 3.146 1.123.345 2.306.535 3.55.535 6.607 0 11.985-5.365 11.985-11.987C23.97 5.39 18.592.026 11.985.026L12.017 0z"
-      />
-    </svg>
-  ),
-  x: (props: IconProps) => (
-    <svg
-      role="img"
-      viewBox="0 0 24 24"
-      xmlns="http://www.w3.org/2000/svg"
-      {...props}
-    >
-      <title>X</title>
-      <path
-        fill="currentColor"
-        d="M18.901 1.153h3.68l-8.04 9.19L24 22.846h-7.406l-5.8-7.584-6.638 7.584H.474l8.6-9.83L0 1.154h7.594l5.243 6.932ZM17.61 20.644h2.039L6.486 3.24H4.298Z"
-      />
-    </svg>
-  ),
-  wear: (props: IconProps) => (
-    <svg
-      {...props}
-      width="87"
-      height="32"
-      viewBox="0 0 84 18"
-      fill="currentColor"
-      aria-describedby="«Rml6»"
-      aria-hidden="true"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <title id="«Rml6»">WEAR</title>
-      <path d="M74.6803 7.59217V3.19634H76.7435C77.8242 3.19634 78.6774 3.34671 79.0271 3.58801C79.7755 4.14055 79.9294 4.70708 79.9294 5.396C79.9294 6.22831 79.6217 6.79134 78.9292 7.21799C78.2893 7.57119 77.5409 7.59217 76.9674 7.59217H74.6768H74.6803ZM76.8869 10.8585L80.384 17.391H84L80.1882 10.159C80.1882 10.159 80.5064 10.0646 80.6043 10.0261C80.8072 9.94221 81.0205 9.86178 81.1499 9.75687C82.6466 8.73222 83.4055 7.26694 83.4055 5.39251C83.4055 3.64746 82.5767 2.11923 81.0135 0.975687C79.8245 0.125895 77.8242 0 76.7365 0H71.0888V17.398H74.6733V10.841H76.8694L76.8869 10.8619L76.8939 10.855L76.8869 10.8585Z"></path>
-      <path d="M54.2013 9.93522L56.1177 4.94488L58.0796 9.99117H54.1803L54.2013 9.93172V9.93872V9.93522ZM59.3351 13.2644L59.3455 13.2924L60.9472 17.398H64.6611L57.9152 0.255287H54.3202L47.5744 17.4015H51.2953L52.9074 13.2679H59.3421L59.3351 13.2609V13.2644Z"></path>
-      <path d="M41.1817 3.44813V0.25179H29.8651V17.398H41.1817V14.1982H33.4531V10.4213H40.8774V7.22848H33.4531V3.44813H41.1817Z"></path>
-      <path d="M19.2829 0.27627L16.0481 10.771L16.0097 10.6451L12.8763 0.27627H10.1276L6.94871 10.771L6.90674 10.6451L3.78385 0.27627L0 0.25179L5.29808 16.1565L5.74571 17.3945H8.2811L8.68676 16.1565L11.5229 6.97319L11.5649 7.10608L14.3625 16.1565L14.7682 17.3945H17.2966L23.0458 0.27627H19.2829Z"></path>
-    </svg>
-  ),
-  zozotown: (props: IconProps) => (
-    <svg
-      {...props}
-      width="94"
-      height="32"
-      viewBox="0 0 189 32"
-      role="img"
-      aria-hidden="true"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path d="M17.173 23.68v4.373H.106V24L11.359 8.373H.479V4h16.373v4L5.759 23.787l11.413-.107zM43.84 16c0 6.88-5.6 12.533-12.48 12.533S18.827 22.986 18.827 16c0-6.88 5.493-12.533 12.533-12.533 6.88 0 12.48 5.653 12.48 12.533m-20.16 0c0 4.267 3.467 7.947 7.627 7.947 4.213 0 7.68-3.68 7.68-7.947 0-4.213-3.413-7.947-7.68-7.947-4.16 0-7.627 3.787-7.627 7.947m38.933 7.68v4.373h-17.12V24L56.746 8.373h-10.88V4h16.373v4L51.146 23.787zM89.173 16c0 6.88-5.6 12.533-12.48 12.533s-12.48-5.6-12.48-12.533c.053-6.88 5.6-12.533 12.587-12.533 6.88 0 12.373 5.653 12.373 12.533m-20 0c0 4.267 3.467 7.947 7.627 7.947 4.213 0 7.68-3.68 7.68-7.947 0-4.213-3.413-7.947-7.68-7.947-4.16 0-7.627 3.787-7.627 7.947zm36.64-7.68h-5.227V28h-4.8V8.32h-5.12V4h15.2zm26.4 7.68c0 6.88-5.6 12.533-12.48 12.533s-12.48-5.6-12.48-12.533c0-6.88 5.493-12.533 12.48-12.533 6.773 0 12.48 5.653 12.48 12.533m-20.16 0c0 4.267 3.467 7.947 7.627 7.947 4.213 0 7.68-3.68 7.68-7.947 0-4.213-3.413-7.947-7.68-7.947-4.16 0-7.627 3.787-7.627 7.947zm44.214 4 5.28-16h5.067l-8.107 22.4-.64 1.653h-3.413l-.587-1.653-4.107-13.333L145.6 26.4l-.587 1.653h-3.467l-.64-1.653L132.799 4h5.12l5.227 16 4.8-16h3.573l4.747 16zm27.253 8.053-9.707-16.427v16.427h-4.8V4h5.12l9.653 16.16V4h4.693v24.053h-4.96z"></path>
-    </svg>
-  ),
-};
