--- src/components/icons/lucide-icons.tsx
@@ -1,39 +1,0 @@
-import {
-  AlertTriangle,
-  ArrowLeft,
-  Calendar,
-  ChevronLeft,
-  ChevronRight,
-  ExternalLink,
-  FileHeart,
-  FileText,
-  Home,
-  Info,
-  Laptop,
-  Mail,
-  Moon,
-  Rss,
-  Sun,
-  Tag,
-  XCircle,
-} from 'lucide-react';
-
-export const Icons = {
-  alertTriangle: AlertTriangle,
-  arrowLeft: ArrowLeft,
-  calendar: Calendar,
-  chevronLeft: ChevronLeft,
-  chevronRight: ChevronRight,
-  externalLink: ExternalLink,
-  fileHeart: FileHeart,
-  fileText: FileText,
-  home: Home,
-  info: Info,
-  laptop: Laptop,
-  mail: Mail,
-  moon: Moon,
-  rss: Rss,
-  sun: Sun,
-  tag: Tag,
-  xCircle: XCircle,
-};
